pyproject.toml
[tool.poetry]
name = "mind-fusion"
version = "0.1.0"
description = ""
authors = ["Sergey <one6sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^30.8.2"
pytest = "^8.3.3"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
fastapi = "^0.115.4"
ruff = "^0.7.3"
uvicorn = "^0.32.0"
punq = "^0.7.0"
pydantic-settings = "^2.6.1"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
